ensure docker is running when trying to run the datadog service
In systemd the `After` directive will only ensure execution ordering of
the specified units if they happen to be started at the same time, but
it does not specify a dependency relation.

`Requires=docker.service` will make sure docker is up and running before
attempting to run the commands of this service.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>